refactor: move execution message type-guards under journal
These type-guards are all journal based and should reside there.
